implemented NetworkConnection

+++ Chat/Chat/Model/UserRemoteRepository.cs
@@ -56,19 +56,25 @@ namespace Chat.Model
         
         public UserRemote GetById(int id)
         {
+            if (id == -1)
+            {
+                return UserRemote.SystemUser;
+            }
+
             // if already fetched serve from memory
             if (_loaded.Any(c => c.Id == id))
             {
                 return _loaded.First(c => c.Id == id);
             }
 
-            List<string[]> result = _dbController.Database.ExecuteSQLQuery("SELECT name, ip, deleted FROM user WHERE id = " + id + ";");
+            List<string[]> result = _dbController.Database.ExecuteSQLQuery("SELECT name, ip, port, deleted FROM user WHERE id = " + id + ";");
 
-            UserRemote userRemote = new UserRemote(bool.Parse(result[0][2]))
+            UserRemote userRemote = new UserRemote(bool.Parse(result[0][3]))
             {
                 Id = id,
                 Name = result[0][0],
-                IP = result[0][1]
+                IP = result[0][1],
+                Port = Int32.Parse(result[0][2])
             };
 
             // store reference
@@ -94,44 +100,53 @@ namespace Chat.Model
         }
         public void Insert(UserRemote obj)
         {
-            _dbController.Database.ExecuteSQLQuery("INSERT INTO user (name, islocal, ip, deleted) VALUES " + "('" + _dbController.Database.Escape(obj.Name) + "', 0, '" + _dbController.Database.Escape(obj.IP) + "', "+(obj.Deleted? 1: 0)+");");
-
-            // set id
-            obj.Id = Int32.Parse(_dbController.Database.LastInsertedId("user"));
-
-            _dbController.Database.ExecuteSQLQuery("INSERT INTO user_has_buddy (userid, buddyid) VALUES " + "(" + obj.BuddyOf.Id + ", " + obj.Id + ");");
-
-            foreach (Conversation conversation in obj.Conversations)
+            if (obj.Id != -1)
             {
-                _dbController.Database.ExecuteSQLQuery("INSERT INTO conversation_has_user (conversationid, userid) VALUES (" + conversation.Id + ", " + obj.Id + ");");
+                _dbController.Database.ExecuteSQLQuery("INSERT INTO user (name, islocal, ip, port, deleted) VALUES " + "('" + _dbController.Database.Escape(obj.Name) + "', 0, '" + _dbController.Database.Escape(obj.IP) + "', " + obj.Port.ToString() + ", " + (obj.Deleted ? 1 : 0) + ");");
+
+                // set id
+                obj.Id = Int32.Parse(_dbController.Database.LastInsertedId("user"));
+
+                _dbController.Database.ExecuteSQLQuery("INSERT INTO user_has_buddy (userid, buddyid) VALUES " + "(" + obj.BuddyOf.Id + ", " + obj.Id + ");");
+
+                foreach (Conversation conversation in obj.Conversations)
+                {
+                    _dbController.Database.ExecuteSQLQuery("INSERT INTO conversation_has_user (conversationid, userid) VALUES (" + conversation.Id + ", " + obj.Id + ");");
+                }
+
+                _bindAutoSaveDelegates(obj);
+
+                // store
+                _loaded.Add(obj);
             }
-
-            _bindAutoSaveDelegates(obj);
-
-            // store
-            _loaded.Add(obj);
         }
 
         public void Remove(UserRemote obj)
         {
-            _loaded.Remove(obj);
-            RemoveById(obj.Id);
+            if (obj.Id != -1)
+            {
+                _loaded.Remove(obj);
+                RemoveById(obj.Id);
+            }
         }
 
         public void RemoveById(int id)
         {
-            _dbController.Database.ExecuteSQLQuery("DELETE FROM user_has_buddy WHERE buddyid = " + id + ";");
-            _dbController.Database.ExecuteSQLQuery("DELETE FROM conversation_has_user WHERE userid = " + id + ";");
-            _dbController.Database.ExecuteSQLQuery("DELETE FROM user where id = " + id + ";");
+            if (id >= 0)
+            {
+                _dbController.Database.ExecuteSQLQuery("DELETE FROM user_has_buddy WHERE buddyid = " + id + ";");
+                _dbController.Database.ExecuteSQLQuery("DELETE FROM conversation_has_user WHERE userid = " + id + ";");
+                _dbController.Database.ExecuteSQLQuery("DELETE FROM user where id = " + id + ";");
+            }
         }
 
         public void Update(UserRemote obj)
         {
-            if (obj != null)
+            if (obj != null && obj.Id != -1)
             {
                 // update activity 
                 _dbController.Database.ExecuteSQLQuery("UPDATE user " +
-                    "SET name = '" + _dbController.Database.Escape(obj.Name) + "', SET ip = '" + _dbController.Database.Escape(obj.IP) + "', deleted = " + (obj.Deleted? 1: 0) + " WHERE id = " + obj.Id + ";");
+                    "SET name = '" + _dbController.Database.Escape(obj.Name) + "', SET ip = '" + _dbController.Database.Escape(obj.IP) + "', port = " + obj.IP.ToString() + ", deleted = " + (obj.Deleted? 1: 0) + " WHERE id = " + obj.Id + ";");
 
                 _dbController.Database.ExecuteSQLQuery("DELETE FROM conversation_has_user WHERE userid = " + obj.Id + ";");
                 foreach (Conversation conversation in obj.Conversations)
